Move all the decoding to the same module
