+++ lib/src/finality.rs
@@ -17,5 +17,6 @@
 
 //! Finality consists is declaring a block as irreversible. It is now forever part of the chain.
 
+pub mod decode;
 pub mod grandpa;
 pub mod justification;
