+++ lib/src/finality/grandpa/commit.rs
@@ -31,5 +31,4 @@
 //! When a commit is received from a third party, it must first be verified. See the [`verify`]
 //! module.
 
-pub mod decode;
 pub mod verify;
