+++ lib/src/sync/all_forks.rs
@@ -84,7 +84,7 @@
 
 use crate::{
     chain::{blocks_tree, chain_information},
-    finality::grandpa,
+    finality::decode,
     header, verify,
 };
 
@@ -988,7 +988,7 @@ impl<TBl, TRq, TSrc> AllForksSync<TBl, TRq, TSrc> {
     ) -> GrandpaCommitMessageOutcome {
         let source = &mut self.inner.blocks[source_id];
 
-        let block_number = match grandpa::commit::decode::decode_grandpa_commit(
+        let block_number = match decode::decode_grandpa_commit(
             &scale_encoded_commit,
             self.chain.block_number_bytes(),
         ) {
